Funciona reservas y paquetes la mitad

+++ src/main/java/co/edu/uniquindio/criterion/model/Habitacion.java
@@ -37,7 +37,7 @@ public class Habitacion implements Serializable {
 
     @NotNull
     @PositiveOrZero
-    private Long precio;
+    private Double precio;
 
     @NotNull
     @ManyToOne
